+++ recipes/default.rb
@@ -18,19 +18,53 @@
 # limitations under the License.
 #
 
-if(node[:nginx][:install_method] == 'source')
-  include_recipe 'nginx::source'
-else
-  include_recipe "nginx::ohai_plugin"
-  if(%w(redhat centos fedora scientific).include?(node[:platform]))
-    include_recipe 'yum::epel'
+include_recipe "nginx::ohai_plugin"
+
+package "nginx"
+
+service "nginx" do
+  supports :status => true, :restart => true, :reload => true
+  action :enable
+end
+
+directory node['nginx']['log_dir'] do
+  mode 0755
+  owner node['nginx']['user']
+  action :create
+end
+
+%w{ nxensite nxdissite }.each do |nxscript|
+  template "/usr/sbin/#{nxscript}" do
+    source "#{nxscript}.erb"
+    mode 0755
+    owner "root"
+    group "root"
   end
-  package "nginx"
-  service "nginx" do
-    supports :status => true, :restart => true, :reload => true
-    action :enable
+end
+
+template "nginx.conf" do
+  path "#{node['nginx']['dir']}/nginx.conf"
+  source "nginx.conf.erb"
+  owner "root"
+  group "root"
+  mode 0644
+  notifies :reload, "service[nginx]"
+end
+
+%w{ sites-enabled sites-available conf.d }.each do |sitedir|
+  directory "#{node['nginx']['dir']}/#{sitedir}" do
+    owner "root"
+    group "root"
+    mode 0755
   end
-  include_recipe 'nginx::commons'
+end
+
+template "#{node['nginx']['dir']}/sites-available/default" do
+  source "default-site.erb"
+  owner "root"
+  group "root"
+  mode 0644
+  notifies :reload, "service[nginx]"
 end
 
 service "nginx" do